In rake tasks, only search Rails engines with defined app/chewy folders
When running the "update all" or "reset all" rake tasks, Chewy would
search all "app/chewy" folders for each Rails engine for indexes to
update or reset. However, if a Rails engine did not have an app/chewy
folder defined, the rake task would fail. This change compacts the
folder list to first remove the list of Rails engines that do not have
the folder defined, before checking to see if there are any existing
indexes in those folders.