Fix routing_missing_exception on delete with parent
Elasticsearch 2 return error when trying to delete an object that must have parent without parent.

This occurs when object is not present, so existing_object is nil. In this case, it was not populating parent key and then causing error routing_missing_exception on Elasticsearch.

The solution is simple not issue a delete if record does not exists.