Improve performance of bypassing
It makes sense to short-circuit Chewy's
`update_proc` when current strategy is `bypass`.

Currently, even when current strategy is to bypass ES,
Chewy still takes time to evaluate the backreference
(which is AR relation, usually).

It's better to skip it as it affects performance of the
application.

Why not put `next if ...` in the beginning of the proc?

Because `Chewy.derive_type` may cause some chewy types to
load, if Rails' `eager_load` is `false`. This can cause
undesired behavior and bugs. So, in order to preserve
existing behavior, we call `derive_type` _before_ `next if`,
and the remaining part after the `next if ...`. The remaining
part is the heaviest one performance-wise, anyway.